requirements.txt
langgraph
langchain-google-genai
langchain
langchain-community
streamlit
python-dotenv
tavily-python
google-generativeai>=0.3.0